+++ src/image.c
@@ -1,25 +1,23 @@
 #include "image.h"
 
-Image image_load(const char *path, int desired_channels) {
-  Image image;
-  stbi_set_flip_vertically_on_load(true);
-  image.data = stbi_load(path, &image.width, &image.height, &image.channels,
-                         desired_channels);
-  return image;
+Image image_load(const char* path, int desired_channels) {
+    Image image;
+    stbi_set_flip_vertically_on_load(true);
+    image.data = stbi_load(path, &image.width, &image.height, &image.channels, desired_channels);
+    return image;
 }
 
-Image image_load_from_memory(const uint8_t *data, int size,
-                                    int desired_channels) {
-  Image image;
-  stbi_set_flip_vertically_on_load(true);
-  image.data = stbi_load_from_memory(data, size, &image.width, &image.height,
-                                     &image.channels, desired_channels);
-  return image;
+Image image_load_from_memory(const uint8_t* data, size_t size, int desired_channels) {
+    Image image;
+    stbi_set_flip_vertically_on_load(true);
+    image.data = stbi_load_from_memory(
+        data, (int)size, &image.width, &image.height, &image.channels, desired_channels);
+    return image;
 }
 
-void image_free(Image *image) {
-  stbi_image_free(image->data);
-  image->data = NULL;
+void image_free(Image* image) {
+    stbi_image_free(image->data);
+    image->data = NULL;
 }
 
-const char *image_get_error_msg(void) { return stbi_failure_reason(); }
+const char* image_get_error_msg(void) { return stbi_failure_reason(); }
